+++ actionpack/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Remove deprecated option `use_route` in controller tests.
+
+    *Rafael Mendonça França*
+
 *   Ensure `append_info_to_payload` is called even if an exception is raised.
 
     Fixes an issue where when an exception is raised in the request the additonal